+++ src/OpenTelemetry.Instrumentation.Runtime/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Updated OTel SDK package version to 1.3.0
   ([#411](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/411))
+* Fix some bugs in Runtime metrics
+  ([#409](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/409))
 * Add GC heap size and refactor GC count as multi-dimensional metrics in Runtime
   metrics ([#412](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/412))